Demo, Date Utils: renames module, renames tests and demo components imports
