+++ tests/e2e/specs/authors/edit.spec.js
@@ -1,9 +1,6 @@
 import Factory from '../../factory'
 import { InitEntityUtils } from '../../lib/commands'
-import {
-  formatDate,
-  parseDate
-} from '../../../../demo/components/authors/authors-date-input-methods'
+import { formatDate, parseDate } from '../../../../demo/utils/dates'
 
 import UI_CONTENT from '../../../../src/constants/ui.content.default'
 import UI_NAMES from '../../../../src/constants/ui.element.names'
